Update access-control.yaml
Adding @bryce-mcmath to the anoncreds-rs (and wrapper) and indy-vdr teams as a member.  Needs to have the CI/CD actions happening automagically when submitting PRs, and he has a number of PRs to do.  @bryce-mcmath is part of the BC Gov team and is a maintainer on OWF Bifold Wallet and the BC Wallet.

Signed-off-by: Stephen Curran <swcurran@gmail.com>

Signed-off-by: Stephen Curran <swcurran@cloudcompass.ca>

Signed-off-by: Stephen Curran <swcurran@gmail.com>